Fixed wheel offsets and navx heading  interpretation
Fixed wheel offsets with inclinometers

The heading was interpreted wrongly by the code at first. When the robot was rotated 90 degrees and I moved the left stick forward, it went backwards. When the robot was rotated by 180 and I moved the left stick forward, it went straight again. This implies that the heading was added while it should have been subtracted from the driving angle. I switched a sign in SwerveDrive.getAngle()'s inputs to fix this. Akilan came up with this.