Logging done with advantage kif
I hope this is it for logging. I used advantageKit, but I specified which directory should be used for logging. This seemed to fix the issue with the robori not being able to find the log file. I got about a dozen log files off the roborio in a row on the rio's internal storag.e